requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
libretranslatepy==2.1.1
lxml==5.1.0
pyaes==1.6.1
pyasn1==0.5.1
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
Telethon==1.34.0
translate==3.6.1
urllib3==2.2.1
